+++ indexed_gzip/zran_file_util.h
@@ -4,8 +4,6 @@
 /*
  * File utilities used to manipulate either
  * Python file-like objects or file descriptors.
- *
- * Author: Paul McCarthy <pauldmccarthy@gmail.com>
  */
 
 #include <stdlib.h>
@@ -15,42 +13,50 @@
 #include <Python.h>
 
 /*
- * Implements a method analogous to fread that is performed on Python file-like objects.
+ * Implements a method analogous to fread that is performed on Python
+ * file-like objects.
  */
 size_t _fread_python(void *ptr, size_t size, size_t nmemb, PyObject *f);
 
 /*
- * Implements a method analogous to ftell that is performed on Python file-like objects.
+ * Implements a method analogous to ftell that is performed on Python
+ * file-like objects.
  */
 int64_t _ftell_python(PyObject *f);
 
 /*
- * Implements a method analogous to fseek that is performed on Python file-like objects.
+ * Implements a method analogous to fseek that is performed on Python
+ * file-like objects.
  */
 int _fseek_python(PyObject *f, int64_t offset, int whence);
 
 /*
- * Implements a method analogous to feof that is performed on Python file-like objects.
+ * Implements a method analogous to feof that is performed on Python file-like
+ * objects.
  */
 int _feof_python(PyObject *f, size_t f_ret);
 
 /*
- * Implements a method analogous to ferror that is performed on Python file-like objects.
+ * Implements a method analogous to ferror that is performed on Python
+ * file-like objects.
  */
 int _ferror_python(PyObject *f);
 
 /*
- * Implements a method analogous to fflush that is performed on Python file-like objects.
+ * Implements a method analogous to fflush that is performed on Python
+ * file-like objects.
  */
 int _fflush_python(PyObject *f);
 
 /*
- * Implements a method analogous to fwrite that is performed on Python file-like objects.
+ * Implements a method analogous to fwrite that is performed on Python
+ * file-like objects.
  */
 size_t _fwrite_python(const void *ptr, size_t size, size_t nmemb, PyObject *f);
 
 /*
- * Implements a method analogous to getc that is performed on Python file-like objects.
+ * Implements a method analogous to getc that is performed on Python
+ * file-like objects.
  */
 int _getc_python(PyObject *f);
 
@@ -87,7 +93,8 @@ int fflush_(FILE *fd, PyObject *f);
 /*
  * Calls fwrite on fd if specified, otherwise the Python-specific method on f.
  */
-size_t fwrite_(const void *ptr, size_t size, size_t nmemb, FILE *fd, PyObject *f);
+size_t fwrite_(
+    const void *ptr, size_t size, size_t nmemb, FILE *fd, PyObject *f);
 
 /*
  * Calls getc on fd if specified, otherwise the Python-specific method on f.
@@ -95,4 +102,4 @@ size_t fwrite_(const void *ptr, size_t size, size_t nmemb, FILE *fd, PyObject *f
 int getc_(FILE *fd, PyObject *f);
 
 
-#endif /* __ZRAN_H__ */
+#endif /* __ZRAN_FILE_UTIL_H__ */